requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.6.0
django-filter==24.2
django-froala-editor==4.2.1
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
PyMySQL==1.1.1
pyotp==2.9.0
python-dotenv==1.0.1
qrcode==8.0
sqlparse==0.5.1
django-unfold==0.55.3 
drf-yasg==1.21.10 